remove broken test
