+++ internal/provider/data_source.go
@@ -22,27 +22,3 @@ func readDataSource(attrObject bw.ObjectType, attrType bw.ItemType) schema.ReadC
 		return objectRead(ctx, d, meta)
 	}
 }
-
-func listOptionsFromData(d *schema.ResourceData) []bw.ListObjectsOption {
-	filters := []bw.ListObjectsOption{}
-
-	filterMap := map[string]bw.ListObjectsOptionGenerator{
-		attributeFilterSearch:         bw.WithSearch,
-		attributeFilterCollectionId:   bw.WithCollectionID,
-		attributeFilterFolderID:       bw.WithFolderID,
-		attributeFilterOrganizationID: bw.WithOrganizationID,
-		attributeFilterURL:            bw.WithUrl,
-	}
-
-	for attribute, optionFunc := range filterMap {
-		v, ok := d.GetOk(attribute)
-		if !ok {
-			continue
-		}
-
-		if v, ok := v.(string); ok && len(v) > 0 {
-			filters = append(filters, optionFunc(v))
-		}
-	}
-	return filters
-}
