+++ internal/provider/data_source_item_login_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestAccDataSourceItemLoginAttributes(t *testing.T) {
@@ -40,32 +39,7 @@ func TestAccDataSourceItemLoginFailsOnInexistentItem(t *testing.T) {
 	})
 }
 
-func TestAccDataSourceItemLoginDeleted(t *testing.T) {
-	var objectID string
-
-	ensureVaultwardenConfigured(t)
-
-	resource.UnitTest(t, resource.TestCase{
-		ProviderFactories: providerFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: tfConfigProvider() + tfConfigResourceItemLoginSmall(),
-				Check:  getObjectID("bitwarden_item_login.foo", &objectID),
-			},
-			{
-				Config: tfConfigProvider() + tfConfigResourceItemLoginSmall() + tfConfigDataItemLoginWithId(objectID),
-				PreConfig: func() {
-					err := bwTestClient(t).DeleteObject("item", objectID)
-					assert.NoError(t, err)
-				},
-				ExpectError: regexp.MustCompile("Error: object not found"),
-			},
-		},
-	})
-}
-
 func TestAccDataSourceItemLoginBySearch(t *testing.T) {
-	var objectID string
 	resourceName := "bitwarden_item_login.foo"
 
 	ensureVaultwardenConfigured(t)
@@ -75,10 +49,7 @@ func TestAccDataSourceItemLoginBySearch(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				Config: tfConfigProvider() + tfConfigResourceFolder() + tfConfigResourceItemLogin(),
-				Check: resource.ComposeTestCheckFunc(
-					checkItemLogin(resourceName),
-					getObjectID(resourceName, &objectID),
-				),
+				Check:  checkItemLogin(resourceName),
 			},
 			{
 				Config: tfConfigProvider() + tfConfigResourceFolder() + tfConfigResourceItemLogin() + tfConfigDataItemLoginWithSearchAndOrg("test-username"),
@@ -98,16 +69,6 @@ func TestAccDataSourceItemLoginBySearch(t *testing.T) {
 			},
 		},
 	})
-}
-
-func tfConfigDataItemLoginWithId(id string) string {
-	return fmt.Sprintf(`
-data "bitwarden_item_login" "foo_data" {
-	provider	= bitwarden
-
-	id 			= "%s"
-}
-`, id)
 }
 
 func tfConfigDataItemLoginWithSearchAndOrg(search string) string {
