+++ internal/provider/object.go
@@ -16,7 +16,7 @@ func objectCreate(ctx context.Context, d *schema.ResourceData, meta interface{})
 }
 
 func objectRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	if _, areFiltersSet := d.GetOk(attributeID); !areFiltersSet {
+	if _, idProvided := d.GetOk(attributeID); !idProvided {
 		return diag.FromErr(objectSearch(d, meta))
 	}
 
@@ -66,6 +66,30 @@ func objectSearch(d *schema.ResourceData, meta interface{}) error {
 	}
 
 	return objectDataFromStruct(d, &obj)
+}
+
+func listOptionsFromData(d *schema.ResourceData) []bw.ListObjectsOption {
+	filters := []bw.ListObjectsOption{}
+
+	filterMap := map[string]bw.ListObjectsOptionGenerator{
+		attributeFilterSearch:         bw.WithSearch,
+		attributeFilterCollectionId:   bw.WithCollectionID,
+		attributeFilterFolderID:       bw.WithFolderID,
+		attributeFilterOrganizationID: bw.WithOrganizationID,
+		attributeFilterURL:            bw.WithUrl,
+	}
+
+	for attribute, optionFunc := range filterMap {
+		v, ok := d.GetOk(attribute)
+		if !ok {
+			continue
+		}
+
+		if v, ok := v.(string); ok && len(v) > 0 {
+			filters = append(filters, optionFunc(v))
+		}
+	}
+	return filters
 }
 
 func objectReadIgnoreMissing(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
